\name{ComradesM}
\alias{ComradesM-package}
\docType{package}
\title{
ComradesM: Runtime and race information data analysed in Chapter 21, "Over 90 years running 90 km", of "Getting (more out of) Graphics"
}

\description{
The runners' times of the Comrades Marathon in South Africa for the years 1921 to 2019 together with race distances over the years and route elevation data.
}

\details{
Most of the datasets analysed in the book are available in GmooG and other R packages.
}

\author{
Antony Unwin unwin@math.uni-augsburg.de

}


